refactor test into 2 files
